+++ data_classes/example/main.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: undefined_class, uri_has_not_been_generated
+
 import 'package:data_classes/data_classes.dart';
 
 part 'main.g.dart';